python/secret_manager_example/requirements.txt
utcp
utcp-http
google-cloud-secret-manager=2.24.0